refactor: move ThemeScript to useTheme hook and remove unused theme utility
