+++ src/components/ui/Waves.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 
 import cn from '@/lib/utils/cn';
 
-const waves = ({ children, className }: { children: React.ReactNode; className?: string }) => {
+const Waves = ({ children, className }: { children: React.ReactNode; className?: string }) => {
     return (
         <div className={cn('relative aspect-square w-full max-w-32', className)}>
             <span className="shadow-raised-sm animate-waves absolute inset-0 h-full w-full rounded-full" />
@@ -13,4 +13,4 @@ const waves = ({ children, className }: { children: React.ReactNode; className?:
     );
 };
 
-export default memo(waves);
+export default memo(Waves);
