+++ src/hooks/useTheme.tsx
@@ -57,4 +57,27 @@ const useTheme = () => {
     return { theme, cycleTheme, nextTheme, setTheme };
 };
 
+export const ThemeScript = () => {
+    return (
+        <script
+            dangerouslySetInnerHTML={{
+                __html: `
+                    (function () {
+                        try {
+                            const theme = localStorage.getItem('theme') || 'system';
+                            const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
+                            const appliedTheme = theme === 'system' ? (prefersDark ? 'dark' : 'light') : theme;
+                            
+                            document.documentElement.setAttribute('data-theme', appliedTheme);
+                            document.documentElement.classList.toggle('dark', appliedTheme === 'dark');
+                        } catch (e) {
+                            console.error("Error applying theme:", e);
+                        }
+                    })();
+                `,
+            }}
+        />
+    );
+};
+
 export default useTheme;
