--- src/lib/utils/theme.tsx
@@ -1,22 +1,0 @@
-export const ThemeScript = () => {
-    return (
-        <script
-            dangerouslySetInnerHTML={{
-                __html: `
-                    (function () {
-                        try {
-                            const theme = localStorage.getItem('theme') || 'system';
-                            const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
-                            const appliedTheme = theme === 'system' ? (prefersDark ? 'dark' : 'light') : theme;
-                            
-                            document.documentElement.setAttribute('data-theme', appliedTheme);
-                            document.documentElement.classList.toggle('dark', appliedTheme === 'dark');
-                        } catch (e) {
-                            console.error("Error applying theme:", e);
-                        }
-                    })();
-                `,
-            }}
-        />
-    );
-};
